requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
pyyaml>=6.0

# Content-based filtering specific
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Development and visualization
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
umap-learn>=0.5.0

# Demo and UI
streamlit>=1.25.0
gradio>=3.40.0

# Testing and quality
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0

# Optional: Experiment tracking
mlflow>=2.5.0
wandb>=0.15.0

# Optional: Fast API for service
fastapi>=0.100.0
uvicorn>=0.23.0
